♻️ reorganize Python bindings
